Add displaying some data

+++ app/src/main/java/khrushchev/ilya/wheatherapp/models/ListWheatherModel.kt
@@ -6,6 +6,6 @@ data class ListWheatherModel(
     @SerializedName("dt")
     val dt: Int,
     val visibility: Int,
-    val pop: Int,
+    val pop: Float,
     val dt_txt: String
 )
